separate navigator and initial list page
